requirements.txt
gitpython
pytest>=3.5.0
python-dateutil
psutil